add integration test for sources w/o auth
