+++ v2/pkg/subscraping/sources/archiveis/archiveis.go
@@ -24,6 +24,11 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	go func() {
 		defer close(results)
 
+		if domain == "" {
+			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Value: "", Error: fmt.Errorf("empty domain error")}
+			return
+		}
+
 		resp, err := session.SimpleGet(ctx, fmt.Sprintf("https://archive.is/*.%s", domain))
 
 		if err != nil {
