+++ v2/pkg/subscraping/sources/crtsh/crtsh.go
@@ -28,6 +28,12 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 
 	go func() {
 		defer close(results)
+
+		if domain == "" {
+			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Value: "", Error: fmt.Errorf("empty domain error")}
+			return
+		}
+
 		count := s.getSubdomainsFromSQL(domain, results)
 		if count > 0 {
 			return
