+++ v2/pkg/subscraping/sources/riddler/riddler.go
@@ -19,6 +19,11 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	go func() {
 		defer close(results)
 
+		if domain == "" {
+			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Value: "", Error: fmt.Errorf("empty domain error")}
+			return
+		}
+
 		resp, err := session.SimpleGet(ctx, fmt.Sprintf("https://riddler.io/search?q=pld:%s&view_type=data_table", domain))
 		if err != nil {
 			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
